Maximally reduce fmpq_mat.h
